Lab #2 - Added input signal value calculation and test for it

+++ Проги/DI.Lab2/DI.Lab2.Base/Exceptions/NonPositiveMessageDurationException.cs
@@ -4,7 +4,7 @@
         : DataInputException
     {
         public NonPositiveMessageDurationException(
-            int duration) : base("Длительность сообщения = " +
+            double duration) : base("Длительность сообщения = " +
                 $"{duration} <= 0")
         { }
     }
